api/requirements.txt
langchain_core
langchain_community
langchain_openai
streamlit
python-dotenv
langserve
fastapi
uvicorn
sse_starlette